all classes changed the map from integer to double in all classes wit…

+++ CoffiiMe/src/main/java/app/coffeeTypes/caffeLatte/CaffeLatte.java
@@ -8,9 +8,9 @@ public class CaffeLatte implements Coffee {
 
     Color color;
     int volume;
-    Map<String, Integer> ingredients;
+    Map<String, Double> ingredients;
 
-    public CaffeLatte(Color color, int volume, Map<String, Integer> ingredients) {
+    public CaffeLatte(Color color, int volume, Map<String, Double> ingredients) {
         this.color = color;
         this.volume = volume;
         this.ingredients = ingredients;
@@ -27,7 +27,7 @@ public class CaffeLatte implements Coffee {
     }
 
     @Override
-    public Map<String, Integer> getIngredients() {
+    public Map<String, Double> getIngredients() {
         return ingredients;
     }
 }
