+++ CoffiiMe/src/main/java/app/coffeeTypes/customCoffee/CustomCoffee.java
@@ -9,9 +9,9 @@ public class CustomCoffee implements Coffee {
 
     Color color;
     int volume;
-    Map<String, Integer> ingredients;
+    Map<String, Double> ingredients;
 
-    public CustomCoffee(Color color, int volume, Map<String, Integer> ingredients) {
+    public CustomCoffee(Color color, int volume, Map<String, Double> ingredients) {
         this.color = color;
         this.volume = volume;
         this.ingredients = ingredients;
@@ -28,7 +28,7 @@ public class CustomCoffee implements Coffee {
     }
 
     @Override
-    public Map<String, Integer> getIngredients() {
+    public Map<String, Double> getIngredients() {
         return ingredients;
     }
 }
